third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-runScript-async-expected.txt

Tests that Runtime.compileScript and Runtime.runScript work with awaitPromise flag.

Running test: testRunAndCompileWithoutAgentEnable
{
    code : 0
    message : Runtime agent is not enabled
}
{
    code : 0
    message : Runtime agent is not enabled
}

Running test: testSyntaxErrorInScript
{
    exceptionDetails : {
        columnNumber : 1
        exception : {
            className : SyntaxError
            description : SyntaxError: Unexpected token }
            objectId : <number>
            subtype : error
            type : object
        }
        exceptionId : 0
        lineNumber : 1
        scriptId : <number>
        text : Uncaught
    }
}

Running test: testSyntaxErrorInEvalInScript
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            className : SyntaxError
            description : SyntaxError: Unexpected token }     at boo.js:2:2
            objectId : <number>
            subtype : error
            type : object
        }
        exceptionId : 0
        lineNumber : 0
        scriptId : <number>
        stackTrace : {
            callFrames : [
                [0] : {
                    columnNumber : 1
                    functionName : 
                    lineNumber : 1
                    scriptId : <number>
                    url : boo.js
                }
            ]
        }
        text : Uncaught
    }
    result : {
        className : SyntaxError
        description : SyntaxError: Unexpected token }     at boo.js:2:2
        objectId : <string>
        subtype : error
        type : object
    }
}

Running test: testRunNotCompiledScript
{
    code : 0
    message : No script with given id
}

Running test: testRunCompiledScriptAfterAgentWasReenabled
{
    code : 0
    message : Runtime agent is not enabled
}
{
    code : 0
    message : No script with given id
}

Running test: testRunScriptWithPreview
{
    result : {
        className : Object
        description : Object
        objectId : <string>
        preview : {
            description : Object
            overflow : false
            properties : [
                [0] : {
                    name : a
                    type : number
                    value : 1
                }
            ]
            type : object
        }
        type : object
    }
}

Running test: testRunScriptReturnByValue
{
    result : {
        type : object
        value : {
            a : 1
        }
    }
}

Running test: testAwaitNotPromise
{
    result : {
        type : object
        value : {
            a : 1
        }
    }
}

Running test: testAwaitResolvedPromise
{
    result : {
        type : object
        value : {
            a : 1
        }
    }
}

Running test: testAwaitRejectedPromise
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            objectId : <number>
            type : object
            value : {
                a : 1
            }
        }
        exceptionId : 0
        lineNumber : 0
        text : Uncaught (in promise)
    }
    result : {
        type : object
        value : {
            a : 1
        }
    }
}
